requirements.txt
watchdog==5.0.3
dropbox==11.36.2
python-dotenv==1.0.1
tenacity==9.0.0
requests==2.32.3
pytest==8.2.2
pytest-mock==3.14.0
